backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.3
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.12
httpx>=0.28.1
geoalchemy2==0.15.2
shapely==2.0.6
psycopg2-binary==2.9.9
python-dotenv==1.0.1
google-genai==1.10.0
email-validator==2.2.0
openpyxl==3.1.5
reportlab==4.2.5
